requirements.txt
beautifulsoup4==4.12.3
lxml==5.2.2
requests==2.32.3
python-dotenv
